+++ lib/errors.ts
