+++ lib/sign.ts
