+++ lib/strings.ts
