+++ lib/verify.ts
