+++ package.json
@@ -1,28 +1,41 @@
 {
   "name": "@ckwalsh/signedsource",
   "version": "0.0.0",
-  "description": "Utilities to sign and verify generated and partially generated files. Inspired by Meta's signedsource package.",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    "import": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    },
-    "require": {
-      "types": "./dist/index.d.cts",
-      "require": "./dist/index.cjs"
-    }
+  "description": "Library and cli to sign and verify generated code. Inspired by Meta's signedsource package.",
+  "homepage": "https://github.com/ckwalsh/signedsource-ts#readme",
+  "bugs": {
+    "url": "https://github.com/ckwalsh/signedsource-ts/issues"
   },
+  "license": "MIT",
+  "author": "Cullen Walsh",
   "files": [
     "dist",
     "README.md",
     "LICENSE"
   ],
+  "type": "module",
+  "exports": {
+    "import": {
+      "types": "./dist/lib/index.d.ts",
+      "import": "./dist/lib/index.js"
+    },
+    "require": {
+      "types": "./dist/lib/index.d.cts",
+      "require": "./dist/lib/index.cjs"
+    }
+  },
+  "module": "./dist/lib/index.js",
+  "types": "./dist/lib/index.d.ts",
+  "main": "./dist/lib/index.cjs",
   "bin": {
     "signedsource": "./dist/bin/signedsource.js"
+  },
+  "directories": {
+    "lib": "dist/lib"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ckwalsh/signedsource.git"
   },
   "scripts": {
     "build": "tsup",
@@ -32,16 +45,6 @@
     "test": "vitest",
     "typecheck": "tsc"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ckwalsh/signedsource-ts.git"
-  },
-  "author": "Cullen Walsh",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ckwalsh/signedsource-ts/issues"
-  },
-  "homepage": "https://github.com/ckwalsh/signedsource-ts#readme",
   "devDependencies": {
     "@arethetypeswrong/cli": "^0.17.3",
     "@eslint/js": "^9.19.0",
